on avance un peu

+++ ft_strlen.c
@@ -6,7 +6,7 @@
 /*   By: ylegzoul <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2019/11/04 11:39:09 by ylegzoul          #+#    #+#             */
-/*   Updated: 2019/11/04 13:32:58 by ylegzoul         ###   ########.fr       */
+/*   Updated: 2019/11/04 14:26:05 by ylegzoul         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,7 +14,7 @@
 
 size_t		ft_strlen(const char *str)
 {
-	int	i;
+	size_t	i;
 
 	i = 0;
 	while (str[i] != '\0')
